# WebApp-phase2/WebApp/specs/001-todo-app/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Todo Application API
  version: 1.0.0
  description: RESTful API for web-based todo application with JWT authentication

servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.example.com/api/v1
    description: Production

components:
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: user@example.com
        created_at:
          type: string
          format: date-time
          example: "2025-12-31T10:00:00Z"
      required:
        - id
        - email
        - created_at

    Task:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        title:
          type: string
          example: "Buy groceries"
        description:
          type: string
          example: "Milk, eggs, bread"
        is_completed:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2025-12-31T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-12-31T10:00:00Z"
      required:
        - id
        - user_id
        - title
        - is_completed
        - created_at
        - updated_at

    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 8
          example: "SecurePassword123"
        password_confirm:
          type: string
          minLength: 8
          example: "SecurePassword123"
      required:
        - email
        - password
        - password_confirm

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          example: "SecurePassword123"
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
        user:
          $ref: "#/components/schemas/User"
      required:
        - access_token
        - token_type
        - user

    TaskCreateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          example: "Buy groceries"
        description:
          type: string
          maxLength: 10000
          example: "Milk, eggs, bread"
      required:
        - title

    TaskUpdateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          example: "Buy groceries"
        description:
          type: string
          maxLength: 10000
          example: "Milk, eggs, bread"
        is_completed:
          type: boolean
          example: true

    TaskListResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        total:
          type: integer
          example: 5
      required:
        - tasks
        - total

    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "User already exists"
      required:
        - detail

paths:
  /auth/register:
    post:
      summary: Register a new user account
      description: Create a new user account with email and password
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input (missing fields, invalid email, password mismatch)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User with this email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      summary: Authenticate user and get JWT token
      description: Login with email and password to receive JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tasks:
    get:
      summary: List all tasks for authenticated user
      description: Retrieve all tasks belonging to the authenticated user
      tags:
        - Tasks
      security:
        - BearerToken: []
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
        "401":
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      summary: Create a new task
      description: Add a new task for the authenticated user
      tags:
        - Tasks
      security:
        - BearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreateRequest"
      responses:
        "201":
          description: Task successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid input (empty title)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tasks/{task_id}:
    get:
      summary: Get a single task
      description: Retrieve details of a specific task (user must own the task)
      tags:
        - Tasks
      security:
        - BearerToken: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Task does not belong to authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      summary: Update a task completely
      description: Replace all updatable fields of a task
      tags:
        - Tasks
      security:
        - BearerToken: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdateRequest"
      responses:
        "200":
          description: Task successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid input (empty title)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Task does not belong to authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      summary: Partially update a task
      description: Update specific fields of a task (title, description, is_completed)
      tags:
        - Tasks
      security:
        - BearerToken: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdateRequest"
      responses:
        "200":
          description: Task successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Task does not belong to authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Delete a task
      description: Remove a task (user must own the task)
      tags:
        - Tasks
      security:
        - BearerToken: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Task successfully deleted
        "401":
          description: Missing or invalid JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Task does not belong to authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"